Add some more debugging bits when replacing kernel
I'm hitting an issue trying to override the kernel in C9S,
and I wanted to trace through things a bit more.

Let's start using the tracing crate more.  One benefit of
this over `g_debug()` is that we avoid calling any of the code
(in this case the creation of the JSON string) unless the log
level is enabled.