Change our failpoint to `return Err` instead of panic
Part of looking at https://github.com/coreos/rpm-ostree/issues/3075

I thought there was an issue with us not catching a panic and
the core dump handler being called before stderr was flushed.

In the end, that wasn't the bug, but I think it's just cleaner
for us to return an error rather than panic.