client: Use `--no-pager` for `systemctl status`
I noticed in gnome-terminal at the default tiny 80x24 size, when
we hit this case `systemctl` will spawn a pager because the status
is too long.  This is confusing and unexpected - I have learned
`systemctl` and `journalctl` may (for the latter, usually) spawn a pager
but I don't expect it from random commands like this.