Add --enable-bin-unit-tests
We lost C unit tests with the Rust-as-main massive buildsystem
rework.

But we're not going to RIIR tomorrow, and there's definitely
a need to continue unit testing directly some of the C/C++
code.

Add a new buildsystem option `--enable-bin-unit-tests` and
turn it on for CI (but not production builds) *or*
when compiling in `--debug` mode.

This basically compiles the unit testing code into the main
binary.  It's hooked up to `make check` too.