core: Support "fileoverride" RPMs
Add support for RPMs which want to override files from other RPMs. This
could be useful generally, but specifically for example in OCP where
we're toying with bundling user-provided overrides as RPMs.

Such RPMs must have the virtual Provides `rpmostree(override)` in order
to make the intent clear and for the logic to kick in.