compose: make "unified core" arguments common
The various subcommands of compose all benefit from knowing
if "unified core" mode is in use. This change makes the
"common" argument set the one that includes the "unified
core" switches and moves the existing set called "common"
(which was common across "install" and "commit" only) to
a more specific name.