extensions: Add support for per-extension repos and modules
While we support arch-specific extensions today, the state of
repos and modules are global.  Which means if a module-using
extension is only available on one architecture (as is the
case in RHEL8/OCP with the `av` module) then it's not
possible to build on other architectures.

The most powerful mechanism here is the `arch-include` design
that we have in the treefile, but...since we already have
an architecture filter per extension, it's easier to add
repos and modules per extension.

Note that the repos and modules are still *global* state, so
we're not supporting actively conflicting extensions right now
still.

Closes: https://github.com/coreos/rpm-ostree/issues/3150