override: Add --cache-only flag
For the same reason we have it elsewhere.  Fetching the repomd
is a fine default, but e.g. someone may want to try overriding
using a local package that doesn't need any dependencies when
the machine is offline, or for some reason the rpm-md repositories
aren't accessible.

Motivated by trying to work around an OpenShift MCO issue where
it only provisions an rpm-md repo transitively.