compose: remove lstchg values from [/usr]/etc/shadow
The format of /etc/shadow contains several fields relating to
password aging. One of these, `lstchg`, contains a timestamp
that represents the epoch day of the last password change. Tools
that don't respect SOURCE_DATE_EPOCH, which includes systemd's
`sysusers` tool, will set this to a value based on the current
time of the build. This causes a lack of reproducibility.

The `lstchg` field can be safely made empty. This disables the
password aging features.

This change rewrites /etc/shadow to remove all `lstchg` values.

This removal could be made conditional on the entry containing
an encrypted password field indicating that the account is either
locked or otherwise restricted from using passwords, such as `*`
or anything starting with `!`.