core: Compare rpm file coloring within layered packages
Before, we were only comparing rpm file coloring between layered
packages and base packages. But we also need to compare layered packages
among themselves. A concrete example is `unixODBC` which isn't in e.g.
Fedora Silverblue, but which some packages pull in both the i686 and
x86_64 versions. In that case, we need to skip the checkout of files in
the i686 version which are also in the x86_64 one.

Closes: #1937