Add `install --force-replacefiles`
Add support for RPMs which want to override files from other RPMs. This
could be useful generally, but specifically for example in OCP where
we're toying with bundling user-provided overrides as RPMs.

We're effectively surfacing rpm's `--replacefiles` switch, though in a
slightly safer way because while we do turn on the associated rpmdb flag
globally (`RPMPROB_FILTER_REPLACEOLDFILES`), we're only allowing the
checkouts of those specific packages to replace files.

This is also rendered separately in `status` as `LocalForcedPackages`.