Unify cliwrap handling into core
This fixes the case of doing cliwrap *and* local package layering,
which ended up wrapping twice, causing an infinite exec loop.

Closes: https://github.com/coreos/rpm-ostree/issues/3184

A lot of history here.  Originally the cliwrap code was duplicated
on the client side (origin files) and the server side (treefiles).
Then lately we've been trying to unify them, which happened via:

https://github.com/coreos/rpm-ostree/pull/2743

I think (but didn't verify) that it was that PR that introduced
this double wrapping.  Basically since we're now passing a synthesized
treefile into the core, it enables cliwrap there too.

However, it's the sysroot upgrader that has the tricky logic right now
around "only invoke libdnf/rpm when necessary" around things like
fetching repodata.  We should drive that into the core too.

For now though, split out the "non-package" bits as a separate API
entrypoint and call it unconditionally from the upgrader, instead
of duplicating the cliwrap logic.