compose: Stabilize --lockfile
We have a lot of experimental features.  Lockfiles have been
in use successfully in FCOS for quite a while now.  While there
are probably things we could improve here, in the end we need
to support what we have today for a long time, so let's just mark
it stable.

Motivated by reducing our number of experimental features (to
pave the way for adding more, i.e. the container bits), as well
as a side conversation I had around supporting skipping depsolve
since that's what osbuild wants.