normalization: Factor out core loop code into function
The normalization code is deeply nested, and I think having
separate functions helps readability and manage complexity.

A key bit in this code motion is we can drop the "bare open block"
(i.e. the `{ ... }` on its own}) which is currently ensuring all
I/O is complete before we recompute the checksum.

Co-authored-by: Colin Walters <walters@verbum.org>