core: Fix minor regression on colorless arches
In 11aeb825d2cb ("core: Handle RPM file colors more efficiently"), we
added an early exit in `handle_file_dispositions` on arches which don't
use colours. However, this is incorrect because that function does more
than just colour handling; it also handles files shared between RPMs.

Fix this and expand the function's docstring.

We have tests for this, but it passed through because CI runs on x86_64.
But this would matter or i686 or armhfp (which is relevant for Fedora
IoT).