status: Ensure lines are sorted and wrapped consistently.
By using the existing print_packages() for
RemovedBasePackages and ReplacedBasePackages multiple package
groups are no longer started in the same line if the next
package/package group would not fit into the same starting line.

A quote parameter was introduced for print_packages() to allow
the unquoted print of package groups (so far used in
RemovedBasePackages and ReplacedBasePackages).

Fix: print_packages() creates a packages_sorted array,
but never actually sorts the package list.

RemovedBasePackages and ReplacedBasePackages do no longer
use the comma separator. Package groups are now either nice
and short (in which case it's okay to print two in the same
line), or new package groups start in a new line -
which obsoletes the comma separator.