countme: Trigger on boot and bi-weekly with random delay
The current timer will trigger weekly with a potential delay of a week,
which means that if the delay is big enough, an entire week will be
skipped and we will miss counting.

Instead, let's trigger on boot and every three days with a random delay
of one day so that we are sure that we at least trigger once per
counting window.

This does not introduce the risk of over counting as the logic for that
check is in rpm-ostree itself.