status: Add version into `status --json`
I wanted to add some code into openshift/machine-config-operator
which checked the rpm-ostree version.  We can do this today with
`rpm-ostree --version` or `rpm -q rpm-ostree`, but the MCO today
already reads `rpm-ostree status --json`.

It's much more natural to parse JSON for this as opposed to forking
the binary again or querying rpm.  (The latter in particular would
require splitting on the upstream version vs release tag etc.)