compose: Add `--disable-repo-locking`
In https://pagure.io/fedora-iot/issue/48
we found a regression from
https://github.com/coreos/rpm-ostree/pull/3193/commits/e0d2a953ff7514063221ab3960f08bb027208228

I think this is really an ostree bug; see
https://github.com/ostreedev/ostree/pull/2493
But for now we need to allow callers to choose either the old or new behavior.
This way e.g. pungi can be configured to lock or not without having
to upgrade/downgrade rpm-ostree.