Start using cap-std/rustix instead of openat/nix
This will be nontrivial churn with no immediate obvious end user
value.  But, long term, the "cap-std ecosystem" of crates has *much*
more momentum than openat.  (Really, it's mostly just Dan Gohman
right now but his code is really, really good; and we can contribute)

There are a few concrete improvements from this; see
https://github.com/bytecodealliance/cap-std#similar-crates
I have to emphasize this bit:

> However, cap-std uses RESOLVE_BENEATH-style resolution where absolute paths are considered errors,

This is a big deal - a while ago I accidentally created an absolute
path in some of the code, and I only noticed it in review.

And related to this, openat's `open()` path does not follow symlinks
by default, which is a huge ergonomic hit when one *does* want to do
so, as I did in a recent patch.

Other bits as noted are that cap-std and the io-lifetimes crates
correctly handles borrowing and ownership for file descriptors, which
is also incredibly important for us.

I just wanted to get the ball rolling here; we can incrementally
port to cap-std/rustix over time.