core: When we fail to find a locked package, also print non-matches
I am hitting a strange issue in coreos-assembler when trying to use
`overrides/rpm`; it generates a lockfile, but then `cosa build`
fails to find the RPM that's clearly there.

Let's print out other packages which match the name and arch.

I don't know what the bug is yet, but this already helped me
debug that the problem is we're somehow missing the overrides
rpm-md repo.