WIP: Directly do container installs
I hit the "no microdnf in testing-devel" problem, and I think
really longer term we need to handle this ourselves and not
fork microdnf.

There's still a longer term where we want to unify of course too.
So I'm indecisive.

Anyways I started on this code which doesn't quite work, and I'm
going to set it aside for now.