rust/treefile: Add client treefile support
Add a new `treefile_new_client` which will be intended to be used in
situations where we're deriving from a base treefile, notably on the
client-side and during container layering.

Enforce that no (currently) "base compose" knobs are set in the
treefile. As we develop, some of these knobs will be promoted to the
higher-level `TreeComposeConfig` struct to support leveraging it
client-side.