rust/treefile: Sort compose knobs into "base" and "derivation"
In some upcoming work, I'll be formalizing a bit more the idea of a
derivation treefile.

To do this sanely, we need to sort out of all the treefile knobs we
support, which ones are valid in only the compose context, or only the
derivation context, or both. We do this by draining all the knobs which
currently are not supported in a derivation context into a new
`BaseComposeConfigFields` struct.

There's a lot of churn from this, but I think the added explicitness is
worth it.