Stabilize `initramfs-etc`
This command has been released for a while now and apart from a few bugs
fixed early on, there hasn't been issues. I'd like to start recommending
it more often, and we have lots of experimental commands, so let's
promote this one.

Keep the old command working for now, but print a deprecation message.