Finish dropping microdnf dependency
The only thing we use it for is clearing the cache. But now that we use
the core, which uses a different cache directory, that command doesn't
delete the right thing. And at that point, it's just simpler to directly
delete it ourselves in the container flow.

Though I think it could also make sense to add a `--no-cache` switch
directly to `rebuild` so that container builders can just apply the
config and clear the cache in one shot.