treefile: Fix parsing of derived config fields
There may be a serde bug going on here.  But I think this is
fallout from https://github.com/coreos/rpm-ostree/pull/3341/commits/72a6a77640c70de0d2a7845b4816727356877bea

Basically we have `#[flatten]` in 3 places; in both the `base`
and `derive` structs, *and* the `base` struct has the `extra` member.

It seems that order matters here.  In order to parse the derive
config fields, it needs to be earlier in the struct.

I am at this moment a bit uncertain why this didn't break other
cases.

But this is prep for supporting `override-remove` in `ex rebuild`
too.