treefile: Use `Deref` trait for something like inheritance
I just changed a single use as a demo, but if we decide
to merge this I can do more of them.  (I want to avoid a conflict
fest though)

This will make it easier to migrate fields between the two structs.

Now, I think what we *really* want here is to flip everything around
here so that we have `DeriveTreefile` and `ComposeTreefile` that
include the base, and deref to that.
But that would be a much larger churn.

The status quo means we can't implement `Deref` to the derive.