Add a `Dockerfile.containerci`
This parallels our cosa build (which does a VM flow) and instead
directly overlays the production FCOS container image.

I am sure we are going to be doing more of this type of thing
as we adapt more of our CI to be really "container native".

My immediate motivation is being able to turn this on in Prow.

(That said...we do want to avoid building our binary twice, because
 it's a bit expensive.  I need to investigate
  https://docs.ci.openshift.org/docs/architecture/ci-operator/#building-artifacts
 perhaps?