composepost: Remove /usr/lib/sysimage/rpm dir before symlinking
Rawhide has moved the rpmdb to `/usr/lib/sysimage/rpm`, which is great!
FCOS composes started breaking because rpm now creates that directory:

https://github.com/coreos/rpm-ostree/issues/3397

We should be able to eventually flip our `_dbpath` macro with this (and
eventually much later once it hits RHEL, stop overriding the macro at
all). This will require some careful changes and testing.

But anyway we probably don't want to do this yet on f35 to match Fedora,
so we will want a transition mechanism; e.g. a `rpmdb-in-sysimage`
treefile knob.

For now just to unblock rawhide changes, let's remove the empty dir so
we can symlink and keep `/usr/share/rpm` as canonical. We'll rework once
we have the permanent solution.