scripts: Use a tmpfs for /run
This is an ugly bug; I noticed while doing an `ostree diff`
between two commits that `/run/ostree-booted` had been added.

This was because bwrap `--ro-bind-data` writes a persistent file,
and we weren't mounting `/run` as a tmpfs.

In general we've been potentially leaking state from any scripts
into the persistent commit!

We should change ostree to perhaps error out on this if `--bootable`
is specified or so.