rust/initramfs: Port to cap-std
This is prep for further cleanup; I want to reduce the number of
APIs we use around child processes.  We right now have Rust `std`
in some places, the `subprocess` crate in this code, and GLib in
others.  This is prep for dropping `subprocess`; the cap-std-ext
code has the file descriptor passing we need.