`db version`: Switch to using SHA-256
Motivated by a RHEL-initiated effort to change or drop code using
SHA-1.

This code is old, has no unit tests, and I am not aware of anything
actually using it.  We could probably just delete it.

That said, let's take a middle ground of changing it to use SHA-256.

I believe the higher level logic here duplicates that of libdnf;
xref https://github.com/rpm-software-management/libdnf/issues/1373
We should be using that API anyways; but that's a larger change,
I'm just trying to get SHA-1 out for now.