treefile: Add conditional includes
The way we use manifests in FCOS and RHCOS really pushes the limits of
rpm-ostree's include system.

In FCOS for example, we often need to have some tweaks applied based on
the version of Fedora being targeted by the stream (especially around
rebase preparations time, and of course rawhide).

Currently, we do this by adding such tweaks in the `manifest.yaml`,
which is unique per stream. But if there are multiple streams that need
this change (e.g. `rawhide`, `branched`, `next-devel`, `next`), it
becomes cumbersome and error-prone to modify all their manifest files.

We could add a conditional include based on `releasever` similar to what
was done with `arch-include`, but this only hints at a need for more
customizable include logic.

Now with the introduction of arbitrary variables, let's add basic
support for a conditional include syntax which should make more
sophisticated schemes possible.

There is a complexity cost here, but overall I think it's well worth it
for the flexibility it affords us in Fedora CoreOS at least.

There's more that could be done after this, such as supporting nested
`TreeComposeConfig` objects instead of having to have a separate file,
but we can leave that as a follow-up.