Fix handling of lockfile parent directory, port to cap_std
openat::Dir::open will fail when asked to open an empty path
(e.g. the result of Path::new("")) which can happen when a compose
command is given a filename with no preceeding path components.

For example, `--ex-write-lockfile-to=./lockfile.json` would
succeed while `--ex-write-lockfile-to=lockfile.json` would fail.

This change special cases the handling of an empty path by
replacing it with `Path::new(".")`.

This change also replaces the use of the openat crate with cap_std.