tree: Remove more try/catch blocks
Some of these can now be removed now that we have macros that fit. The
higher-level try/catch block in `rpmostree-compose-builtin-tree.cxx` we
can remove because I've now verified that we're consistently handling
exceptions at every throwable callsite in the codebase.

Leave the one in `rpmostreed-transaction.cxx` so that in case we regress
on consistently wrapping CXX calls, the daemon can still recover. But
print a scary error in the journal to track this.