treefile: rework validation for container flow
In preparations for enabling some of the derivation knobs for the
container workflow, let's drop the `error_if_nonempty` bit from
`treefile_new_client_from_etc()`.

Instead, add a separate `validate_for_container` that we call from the
container rebuild method.

While we're here, add a bunch of asserts in the core that also verifies
only supported knobs are being used in the container flow.