app/override: Support override replace in container flow
This one requires a bit more work. The main issue is that replacement
overrides usually heavily leverage the package cache repo. But here we
don't have that.

I think it's simpler to keep them separate instead of trying to make the
same key work in both modes. So we add a new
`override-replace-local-rpms` which takes paths to local RPMs.

Then in the core, it's a matter of adding the packages to the sack and
the goal just like we do for cached packages.

Note this only works for trivial packages for now. E.g. replacing the
kernel with this does not work yet. We also need support for fetching
overrides from repos.