vmcheck/history: delete journal hack
This is a racy hack because the journal could still be flushing things
while we try to delete it.

But anyway, it was needed when we would reuse the same VM for multiple
vmcheck tests. This is no longer the case since rebasing on top of
`kola`. So just nuke it.