compose extensions: Support `--rootfs`
As we're working on "container native" (rpm)-ostree, it feels
natural to have the extensions build happen as part of e.g. a
`Dockerfile` build that simply pulls in the base image.

And to do that, we don't need any of this "offline" stuff that
the compose extensions path is using.  Here we're running
rpm-ostree inside a container, and conceptually we mostly just
need the libdnf side.