ci: Add Github action flow for testing container path
I was originally going to do this in Prow, but it is less painful
in GH actions.  Drop the `Dockerfile` for this that isn't used,
and move the script into the main `ci/` dir.

This is modeled loosely after the code in https://github.com/ostreedev/ostree-rs-ext/blob/main/.github/workflows/rust.yml