rust/origin: support converting treefile to origin
I'd like to implement https://github.com/coreos/rpm-ostree/issues/1265
for both the client-side and the container flow. But to do that, I'd
rather not have to wire it through all the layers that need to know
about it.

In a move towards https://github.com/coreos/rpm-ostree/issues/2326, I'd
like to try to have everything in the `DeployTransaction` ->
`RpmOstreeSysrootUpgrader` -> `RpmOstreeContext` flow only deal in
treefiles. But to not change the format on disk just yet, we still need
to deserialize from and serialize back into a `GKeyFile` at the I/O
boundaries (really, libostree API boundaries).

We have a function for the former, but not the latter. This patch adds
one.