libpriv/origin: Move source of truth to treefile
This makes the source of truth of `RpmOstreeOrigin` a `Treefile`. The
`GKeyFile` is now considered a "second layer" cache. This is mostly
symbolic for now but it paves the way for slowly converting methods over
to use the `Treefile` directly. Until eventually, `RpmOstreeOrigin` will
just be a thin wrapper around `Treefile`, at which point we can cut over
and have callers use the `Treefile` API directly and delete
`RpmOstreeOrigin`.

For every setter, sync any changes back to the treefile by recreating
it via `origin_to_treefile`.