libpriv/origin: Stop commenting override commit version
We won't be able to retain this as we move to a treefile world. We could
make it a new informational field, but that can be confusing and I don't
think it's very necessary anyway since `ostree admin status`/
`rpm-ostree status` (or `ostree show` if both of those are broken) will
still show the version.