daemon: Handle `org.freedesktop.DBus.Error.ServiceUnknown` too
For some reason on at least current C9S, dbus-broker is returning
a different error when polkit is not available.  Looking at it,
the two errors come from here
https://github.com/bus1/dbus-broker/blob/92142f321fc80c76ab1cc973ce2defa1039cac4a/src/bus/driver.c#L743

But I don't quite understand the circumstances in which one is sent
versus the other.

In any case, this should be a very safe and targeted change to
ensure we handle both errors.

Closes: https://github.com/coreos/rpm-ostree/issues/3554