compose: Fix handling of base rev
I definitely tested the previous patch (and so did Joseph)
but I am fearing that I did further cleanup before pushing
that broke it.

What we're seeing right now is a segfault trying to use this with
RHCOS (which notably doesn't have a `ref` in the treefile).

Anyways, we should only be trying to resolve the base rev in
the "outside" path, not the in-container path.  (Right now, but
once we can depend on a newer ostree inside the container we can
actually make this all work symmetrically)