use basename when cleaning up the .vmlinuz.hmac file
Prior to this commit, the .vmlinuz.hmac file was assumed to be:

<hmac>  /boot/vmlinuz-<version>

This is because the ARK kernel.spec so-happens to run hmac512sum on that path while building the RPM. Because absolute paths to the kernel don't make sense for Core-OS, `/boot/` was removed from the start of the path. This means that the prior emergent behavior was a `basename`, where supported. Unfortunately, some custom kernels build with a path other than `/boot/`, and the prior behavior would log an error and abort.

At the time of this commit, using the basename is under consideration in ARK. However, there are many existing custom kernels that rpm-ostree users may want to use.

Given that a basename was being performed (when no error is encountered), this change reflects that behavior. While this is still a hack, it is likely a more robust hack.

Two spaces are expected to delimit the hash and the path.