daemon: Move tokio handle reference into daemon
We got a panic inside `rpmostreed_transaction_new_cleanup` because
there was no tokio handle associated with the current thread.

The goal of the original PR here was to grab the tokio context
from the main (initial) thread.  Now, in that stack trace
GDBus is clearly executing the callback on a worker thread.

As best I can tell from looking at the GDBus source, this can
happen in cases where method authorization is in play; there's a
whole dance where that bit gets dispatched to a worker, then we
try to schedule the callback on the original main context.

(Also, one can configure all GDBus method invocations to run
 in a thread, but we're not doing that)

I am not totally sure why this isn't *always* crashing - it might
only occur in situations where the main context is not idle?

Anyways, I think it will be more reliable instead to have the C++
global daemon context hold a reference to the tokio handle.  All
our GDBus stuff on the daemon side uses that, and we know that
object is constructed on the `tokio::task::spawn_blocking` thread
created by tokio, not a GLib-created thread.