ci: Move legacy compose tests to GH actions
The compose tests keep flaking out in FCOS CI/Jenkins.  I am not
totally sure why, but I think it's related to the fact it's on
bare metal (which is great) but that Kubernetes uses CPU shares
and most tools which try to count processors are confused by that.

We have 3 CI systems, and actually I think GH actions is a good
match for our legacy compose tests.  It gives us a VM, in which
we can easily run a privileged container.

Further, GH actions is nicer here in that the legacy compose tests
can appear as a separate check status.