rust/treefile: Add container_image_reference derivation field
Having the container case serialize to the `base_refspec` field is
weird. But also, from a YAML perspective we want users to type a
different field for the container flow.

This also fixes a subtle issue where transforming from treefile back to
origin would not use the correct field for the container case.