rust/treefile: Make package and modules-related fields BTreeSets
Both for reproducibility, but also because it'll be easier to implement
change detection in the future when migrating origin logic.

In the process, also convert `ignore_removed_users` and
`ignore_removed_groups` over to `BTreeSet`.

There are no treefile fields now which use `HashSet`. It's possible
there are more `Vec` fields though which should be `BTreeSet`s.