Add support for repo overrides in container flow
This implements the semantics discussed in
https://github.com/coreos/rpm-ostree/issues/1265.

For now, it's only supported in treefile format. Will wire it up to
`override replace` soon.

Also for now, this only works in the container flow. The client-side
will come soon, but it still needs more cleanups there to avoid having
to wire this through a non-trivial number of APIs.