tree: Fix formatting for GLNX_HASH_TABLE_FOREACH
In the clang reformatting change, we lost indentation of the bodies for
`GLNX_HASH_TABLE_FOREACH` macros because `clang-format` thinks it's a
function call-style macro. This can lead to misreading the code and
possibly cause bugs.

Thankfully, it has an option for declaring foreach-style macros. Use it
and reformat the whole tree.