main: Move `ostree-container` to pkglibdir, also expose `ostree ima-sign`
Quite a while ago we merged https://github.com/ostreedev/ostree/pull/2515
but I forgot to follow up here to make use of it once the ostree
release had been done.  It's cleaner to avoid the "namespace pollution"
of having `/usr/bin/ostree-container`; instead it is canonically
`ostree container`.

Next, we have code in ostree-ext for IMA signing. The CLI wrapping code
here is fully generic already, rename it to make that clearer, and
add the `ostree-ima-sign` case.