cliwrap: Enable `yum install foo` in a container
For the "ostree native container" bits, the goal is to make things
*extremely obvious* to someone who is using a `Dockerfile` to build
derived images.

Another way to say this is that a goal is for ostree to
"fade into the background".  It's still there - just not something
everyone needs to learn or care about.

This makes it so that typing `yum install usbguard && yum clean all`
Just Works as one would expect.