postprocess: Automatically propagate user.ostreemeta xattrs in commit
This is actually what we should have been doing from the start
for all metadata actually.  It is a key to ultimately fully operating
as non-root.

This is specifically prep for having the importer pull IMA signatures
from RPM headers, translate them into the standard `security.ima`
value, but store them inside a `bare-user` repo in the `user.ostreemeta`
xattr.