Enable compilation with librhsm
First, add a generic `--enable-featuresrs=$x` autoconf bit which just passes
through to `cargo build --features $x` so we don't need to mirror all
future options in autoconf.

Use this to add `--enable-featuresrs=rhsm` which basically boils
down to passing `-DENABLE_RHSM_SUPPORT=1` to the libdnf cmake bits.

This is important for subscriptions on RHEL.