tests: Fix dependency on uid/gid in compose tests
Trying to do a build in GH actions, I was really confused initially
why `test_tmpfiles_d_translation()` was failing there.  It turns
out the test was indirectly depending on the uid being 4 characters
long (1000, from coreos-assembler) because we end up writing the
current uid into the filesystem, and then the test asserts on the
total size of all regular files.

Fix the test to dynamically compute the expected size.