rust/libdnf-sys: Bridge using C++ classes instead
Instead of having individual separate functions acting on a pointer to a
libdnf object, bundle them up into proper classes instead. This not only
makes it feel cleaner on the Rust side, but also has the major advantage
of being able to hold a reference to an object. This means that it's
easier to wrap libdnf functions that return new references for use on
the Rust side, which is something we'll do in an upcoming patch.