treefile: Drop override-replace-local-rpms
The addition of `override-replace-local-rpms` was a hack to bridge the
gap between the container and client-side flow where only the latter has
access to a pkgcache repo.

This patch enhances the existing `override-replace-local` key to work in
the container flow: we stage RPMs in a directory in `/run`, and the core
knows to look there for them.

This then allows us to remove the `override-replace-local-rpms` key and
clean up a bunch of things related to it.

To be nice, we might have to in the future expose the staging logic as a
separate command which outputs on stdout the digest and NEVRA so that
users can include them in their YAML dropins.

This is prep also for supporting local RPMs, which will work a similar
way.