rust/fedora-integration: Only download valid packages to replace
Right now, when applying overrides from Bodhi or Koji, we just download
all the packages, regardless of whether they're valid replacements or
not. This leads to lots of wasted I/O and inactive overrides.

Enhance things so that override replacements are now filtered down so
that only RPMs that are applicable to the base are downloaded. Implement
this for both the client-side and container paths.