rust/treefile: Add `merge_treefile`
I'd like to enhance the `UpdateDeployment` D-Bus API to support passing
through a treefile string. Going forward, this should replace the need
for adding any other individual keys to the modifiers parameter as new
features get added. Instead, everything should be focused around
treefile dropins and the CLI thought of as an imperative interface over
declarative "dynamic dropins".

To support this though, we need to provide a way for the daemon to merge
into the current treefile these new treefile dropins.