rust/scripts: Ignore `kernel-automotive-core.posttrans`
Yet another kernel package variant whose %posttrans we want to ignore.

As mentioned in there, I think we should change the strategy so that we
detect kernel packages like we do in the core based on its `Provides` to
know to ignore %posttrans.

Alternatively/additionally, we should work with kernel and systemd teams
so that e.g. it short-circuits if it detects an "offline" system. That
would also benefit containers, which AFAIK even today runs dracut on
kernel installs.

Closes: #3720